requirements.txt
opencv-python
tqdm
pyyaml
imageio
imagecodecs
torch 
torchvision 
pytorch_lightning